requirements.txt
click>8.1.3
gpsd-py3>0.3.0
